Create index.html
Bulma Registration Form
In this Bulma Form, we used the grid system of this framework and also used new helper classes that come in new version 0.9. Here we modify the Bulma framework class and also used custom background color. This form is fully responsive to all views.

Make it yours now by using it, downloading it, and please share it. we will design more elements for you.